Added docstrings to Parking model and services; Changed ParkingsService to ParkingService, getAll to list, getOne to retrieve, getOneExtended to retrieveExtended; Updated path and file name from parkings to parking
