+++ web/src/app/shared/services/parking/parking.service.spec.ts
@@ -1,13 +1,13 @@
 import { TestBed } from '@angular/core/testing';
 
-import { ParkingsService } from './parkings.service';
+import { ParkingService } from './parking.service';
 
-describe('ParkingsService', () => {
-  let service: ParkingsService;
+describe('ParkingService', () => {
+  let service: ParkingService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({});
-    service = TestBed.inject(ParkingsService);
+    service = TestBed.inject(ParkingService);
   });
 
   it('should be created', () => {
